Clock: fix layout for 320dp devices
800x480px dpi devices have less screen space than the xhdpi devices
Google designed the layout for and the layout needs adjustment.

The number in the timer (left-most tab) is cut off on older devices.
Adjust the weight (per Jie W.'s suggestion) so that it can be
displayed properly.

It is impossible to account for all languages with very long translation
for "H" "M" and "S", and that can only be fixed by de-localization or
re-localization.

This should not have any effect on devices with larger screens.

Change-Id: If61b676f220a6f1c39105fe3948946100a5f64b2