DeskClock: Add increasing volume option for alarm clocks
this commit adds a checkbox preference for "increasing volume" to
alarm clocks. when checked, the alarm will start off at 10% of the
system-wide alarm volume and slowly increase to 100% over the
course of a minute.

Original author for jb is Austen Dicken<cvpcsm@gmail.com>
jb-mr1 porters Davor Bertovic (LorD ClockaN) and maxwen

Change-Id: Ic097e344ed1570602f17f774faed7b612d47be73